# exports/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tickets/validate-purchase": {
      "post": {
        "description": "Step 1: Validates ticket availability, sales window, and calculates total price. Call this before creating payment. Does not create tickets.",
        "operationId": "TicketController_validatePurchase",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Purchase validated. Proceed to payment.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "event": {
                      "_id": "...",
                      "title": "Concert"
                    },
                    "tier": {
                      "name": "VIP",
                      "price": 15000
                    },
                    "subtotal": 30000,
                    "currency": "NGN",
                    "quantity": 2
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (sold out, sales ended, etc.)"
          },
          "404": {
            "description": "Event or tier not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Validate ticket purchase before payment",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/tickets": {
      "get": {
        "description": "Returns all tickets purchased by the user with event details",
        "operationId": "TicketController_getUserTickets",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter tickets by status",
            "schema": {
              "type": "string",
              "enum": [
                "valid",
                "used",
                "cancelled",
                "refunded"
              ]
            }
          },
          {
            "name": "eventId",
            "required": false,
            "in": "query",
            "description": "Filter tickets for a specific event",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          },
          {
            "name": "upcoming",
            "required": false,
            "in": "query",
            "description": "Show only tickets for upcoming events",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "past",
            "required": false,
            "in": "query",
            "description": "Show only tickets for past events",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Pagination page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Results per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tickets",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "tickets": [
                      {
                        "_id": "...",
                        "ticketNumber": "TKT-2025-ABC12",
                        "event": {
                          "title": "Concert"
                        },
                        "status": "valid",
                        "qrCode": {
                          "data": "...",
                          "generatedAt": "..."
                        }
                      }
                    ],
                    "pagination": {
                      "total": 10,
                      "page": 1,
                      "limit": 20,
                      "totalPages": 1
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get all tickets for the authenticated user",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/tickets/{id}": {
      "get": {
        "description": "Returns full ticket information including QR code",
        "operationId": "TicketController_getTicketById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MongoDB ObjectId of the ticket",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket details",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "_id": "...",
                    "ticketNumber": "TKT-2025-ABC12",
                    "event": {
                      "title": "Concert"
                    },
                    "status": "valid",
                    "qrCode": {
                      "data": "...",
                      "generatedAt": "..."
                    },
                    "attendeeName": "John Doe",
                    "attendeeEmail": "john@example.com"
                  }
                }
              }
            }
          },
          "403": {
            "description": "You do not own this ticket"
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get ticket details by ID",
        "tags": [
          "Tickets"
        ]
      },
      "delete": {
        "description": "Cancel a ticket at least 24 hours before the event. Ticket status will be updated and refund will be processed.",
        "operationId": "TicketController_cancelTicket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MongoDB ObjectId of the ticket",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket cancelled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Ticket cancelled successfully. Refund will be processed.",
                    "ticket": {
                      "_id": "...",
                      "status": "cancelled"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Cannot cancel (within 24h of event or already scanned)"
          },
          "403": {
            "description": "You do not own this ticket"
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Cancel ticket and request refund",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/tickets/{id}/qr-code": {
      "get": {
        "description": "Returns QR code as base64 data URL that can be displayed or printed",
        "operationId": "TicketController_getTicketQRCode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MongoDB ObjectId of the ticket",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "QR code image",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "ticketNumber": "TKT-2025-ABC12",
                    "qrCodeImage": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
                  }
                }
              }
            }
          },
          "403": {
            "description": "You do not own this ticket"
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get QR code image for a ticket",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/tickets/scan": {
      "post": {
        "description": "Creator scans attendee QR code to verify and admit entry. Only the event creator can scan tickets for their events.",
        "operationId": "TicketController_scanTicket",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scan result",
            "content": {
              "application/json": {
                "schema": {
                  "examples": {
                    "success": {
                      "value": {
                        "success": true,
                        "message": "Ticket scanned successfully",
                        "ticket": {
                          "ticketNumber": "TKT-2025-ABC12"
                        },
                        "scannedAt": "2025-08-15T19:30:00Z"
                      }
                    },
                    "alreadyScanned": {
                      "value": {
                        "success": false,
                        "message": "Ticket has already been scanned",
                        "scannedAt": "2025-08-15T19:15:00Z",
                        "ticket": {
                          "status": "used"
                        }
                      }
                    },
                    "invalidStatus": {
                      "value": {
                        "success": false,
                        "message": "Ticket is cancelled",
                        "ticket": {
                          "status": "cancelled"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid QR code"
          },
          "403": {
            "description": "Only the event creator can scan tickets"
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Scan ticket QR code at event entrance",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/tickets/event/{eventId}": {
      "get": {
        "description": "Returns all tickets sold for the event with buyer information. Only event creator can access.",
        "operationId": "TicketController_getEventTickets",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "description": "MongoDB ObjectId of the event",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event tickets and statistics",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "tickets": [
                      {
                        "ticketNumber": "TKT-2025-ABC12"
                      }
                    ],
                    "stats": {
                      "total": 150,
                      "valid": 120,
                      "used": 25,
                      "cancelled": 3,
                      "refunded": 2,
                      "scanned": 25,
                      "byTier": {
                        "507f191e810c19729de860ea": 100,
                        "507f191e810c19729de860eb": 50
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "You do not own this event"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get all tickets for an event",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/tickets/event/{eventId}/stats": {
      "get": {
        "description": "Returns aggregated statistics about ticket sales and scanning for the event.",
        "operationId": "TicketController_getTicketStats",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "description": "MongoDB ObjectId of the event",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket statistics",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "overall": {
                      "total": 150,
                      "valid": 120,
                      "used": 25,
                      "cancelled": 3,
                      "refunded": 2,
                      "scanned": 25
                    },
                    "byTier": [
                      {
                        "_id": "507f191e810c19729de860ea",
                        "count": 100,
                        "scanned": 15
                      },
                      {
                        "_id": "507f191e810c19729de860eb",
                        "count": 50,
                        "scanned": 10
                      }
                    ]
                  }
                }
              }
            }
          },
          "403": {
            "description": "You do not own this event"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get ticket statistics for an event",
        "tags": [
          "Tickets"
        ]
      }
    },
...
  },
  "info": {
    "title": "Events API",
    "description": "API for creating and managing events",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AttendeeInfoDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the attendee for this ticket",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "Email address of the attendee",
            "example": "john.doe@example.com"
          },
          "phone": {
            "type": "string",
            "description": "Phone number in Nigerian format",
            "example": "08012345678"
          }
        }
      },
      "PurchaseTicketDto": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "description": "MongoDB ObjectId of the event",
            "example": "507f1f77bcf86cd799439011"
          },
          "tierId": {
            "type": "string",
            "description": "MongoDB ObjectId of the ticket tier",
            "example": "507f191e810c19729de860ea"
          },
          "quantity": {
            "type": "number",
            "description": "Number of tickets to purchase. Maximum 10 per transaction.",
            "example": 2,
            "minimum": 1,
            "maximum": 10
          },
          "attendees": {
            "description": "Optional attendee information for each ticket. If provided, array length must match quantity. If purchasing for yourself only, you can omit this field.",
            "example": [
              {
                "name": "John Doe",
                "email": "john@example.com",
                "phone": "08012345678"
              },
              {
                "name": "Jane Smith",
                "email": "jane@example.com",
                "phone": "08087654321"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttendeeInfoDto"
            }
          }
        },
        "required": [
          "eventId",
          "tierId",
          "quantity"
        ]
      },
      "ScanTicketDto": {
        "type": "object",
        "properties": {
          "qrCodeData": {
            "type": "string",
            "description": "Encrypted QR code data from the ticket",
            "example": "U2FsdGVkX1+5h3j2k4l6m8n9o0p1q2r3s4t5u6v7w8x9y0z1a2b3c4d5"
          }
        },
        "required": [
          "qrCodeData"
        ]
      },
...
    }
  }
}